pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "levy"
version = "0.1.0"
description = "One decorator to monetize any FastAPI endpoint via Tempo/MPP."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pympp[tempo]>=0.5",
    "fastapi>=0.115",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "uvicorn>=0.30",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["levy"]
